backend/pyproject.toml
[project]
name = "multi-user-todo-backend"
version = "0.1.0"
description = "Backend for the Multi-User Todo Web Application"
authors = ["Developer <developer@example.com>"]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlmodel>=0.0.16",
    "psycopg2-binary>=2.9.9",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "alembic>=1.13.1",
    "pytest>=7.4.3",
    "httpx>=0.25.2",
    "resend>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.12.1",
    "flake8>=6.1.0",
    "mypy>=1.8.0",
    "isort>=5.13.2",
]